updated data preparation scripts and added prediction scripts

+++ GEFA/prepare_data.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-source ~/miniconda3/etc/profile.d/conda.sh
+source /home/azureuser/miniconda3/etc/profile.d/conda.sh # change this to match the location of your conda.sh file
 
 cmd=`readlink -f $0`
 cmdDir=`dirname $cmd`
@@ -19,7 +19,7 @@ then
     echo "Usage: $0 dataset_name"
     exit 1
 fi
-conda activate RaptorX
+conda activate RaptorX # a conda environment named RaptorX with all the required packages for running RaptorX is required to run this script
 # Gener
 for seq in ${cmdDir}/data/${1}/sequences/*.fasta
 do
@@ -30,31 +30,36 @@ do
         mkdir ${map_dir}/${name}
     fi
 
-    ${DistFeatureHome}/BuildMSAs.sh -d ${map_dir}/${name} $seq
-    ${DistFeatureHome}/GenDistFeaturesFromMSA.sh -o ${map_dir}/${name} ${map_dir}/${name}/${name}_OUT/${name}_contact/${name}_uce5/${name}.a3m
-    ${DL4DistancePredHome}/Scripts/PredictPairRelation4OneInput.sh -d ${map_dir}/${name} ${map_dir}/${name}/${name}.inputFeatures.pkl
-    ${DL4DistancePredHome}/Scripts/PrintContactPrediction.sh ${map_dir}/${name}/${name}.predictedDistMatrix.pkl ${map_dir}/${name}
-
-    python txt2npy.py ${map_dir}/${name}/${name}.CM.txt
-    mv ${map_dir}/${name}/${name}.npy ${map_dir}/${name}.npy
-    rm -r ${map_dir}/${name}
+    if [ ! -f ${map_dir}/${name}.npy ]
+    then
+        ${DistFeatureHome}/BuildMSAs.sh -d ${map_dir}/${name} $seq
+        ${DistFeatureHome}/GenDistFeaturesFromMSA.sh -o ${map_dir}/${name} ${map_dir}/${name}/${name}_OUT/${name}_contact/${name}_uce5/${name}.a3m
+        ${DL4DistancePredHome}/Scripts/PredictPairRelation4OneInput.sh -d ${map_dir}/${name} ${map_dir}/${name}/${name}.inputFeatures.pkl
+        ${DL4DistancePredHome}/Scripts/PrintContactPrediction.sh ${map_dir}/${name}/${name}.predictedDistMatrix.pkl ${map_dir}/${name}
+        python txt2npy.py ${map_dir}/${name}/${name}.CM.txt
+        mv ${map_dir}/${name}/${name}.npy ${map_dir}/${name}.npy
+        rm -r ${map_dir}/${name}
+    fi
     
     if [ ! -d ${profile_dir}/${name}_PROP ]
     then
         mkdir ${profile_dir}/${name}_PROP
     fi
+
     cd ${repoDir}/Predict_Property
-    . ${repoDir}/Predict_Property/Predict_Property.sh -o ${profile_dir}/${name}_PROP -i $seq
+    ${repoDir}/Predict_Property/Predict_Property.sh -o ${profile_dir}/${name}_PROP -i $seq
     cd ${cmdDir}
 done
 
-conda activate tape
+conda activate tape # a conda environment named tape with tape proteins library installed is required
 for seq in ${cmdDir}/data/${1}/sequences/*.fasta
 do
     name=`basename $seq .fasta`
-    
-    tape-embed unirep $seq ${cmdDir}/data/${1}/emb/${name}.npz babbler-1900 --tokenizer unirep --full_sequence_embed
+    if [ ! -f ${cmdDir}/data/${1}/emb/${name}.npz ]
+    then
+        tape-embed transformer $seq ${cmdDir}/data/${1}/emb/${name}.npz bert-base --tokenizer unirep --full_sequence_embed
+    fi
 done
 
-conda activate geometric
+conda activate geometric # a conda environment named geometric with libraries required to run GEFA model
 python prepare_data.py $1
